changed model: many attachment content for one attachment name; add prefix for internal wiki pages

+++ elwikijsp.root/bundles_elwiki/org.elwiki.data.persist/src/org/elwiki/data/persist/json/deserialize/PageContentDeserializer.java
@@ -21,7 +21,7 @@ public class PageContentDeserializer extends DeserialiseStuff implements JsonDes
 
 		PageContent pageContent = Elwiki_dataFactory.eINSTANCE.createPageContent();
 
-		int version = getInt(jsonObject, "version");
+		short version = (short) getInt(jsonObject, "version");
 		String lastModify = getString(jsonObject, "lastModify");
 		String author = getString(jsonObject, "author");
 		String changeNote = getString(jsonObject, "changeNote");
